moved site classes to calc package
