+++ src/org/opensha2/calc/CalcConfig.java
@@ -21,7 +21,6 @@ import org.opensha2.eq.model.SourceType;
 import org.opensha2.gmm.Gmm;
 import org.opensha2.gmm.GroundMotionModel;
 import org.opensha2.gmm.Imt;
-import org.opensha2.util.Site;
 
 import com.google.common.base.Optional;
 import com.google.common.collect.Maps;
