+++ src/org/opensha2/calc/Deaggregator.java
@@ -18,7 +18,6 @@ import org.opensha2.geo.Location;
 import org.opensha2.geo.Locations;
 import org.opensha2.gmm.Gmm;
 import org.opensha2.gmm.Imt;
-import org.opensha2.util.Site;
 
 import com.google.common.base.Function;
 import com.google.common.collect.ImmutableMap;
