+++ src/org/opensha2/gmm/GmmInput.java
@@ -19,11 +19,11 @@ import static org.opensha2.gmm.GmmInput.Field.ZHYP;
 import static org.opensha2.gmm.GmmInput.Field.ZTOP;
 import static org.opensha2.internal.TextUtils.NEWLINE;
 
+import org.opensha2.calc.Site;
 import org.opensha2.eq.Magnitudes;
 import org.opensha2.eq.fault.Faults;
 import org.opensha2.eq.model.Distance;
 import org.opensha2.eq.model.Rupture;
-import org.opensha2.util.Site;
 
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Optional;
